Split in 2 as 37 tests in a single file is a bit on the high side.
